split as number anonymization tests into separate file
